requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
astunparse==1.6.3
attrs==23.2.0
certifi==2023.11.17
cffi==1.16.0
colorama==0.4.6
docutils==0.20.1
exceptiongroup==1.2.0
h11==0.14.0
idna==3.6
importlib-metadata==7.0.1
iniconfig==2.0.0
MarkupSafe==2.1.3
outcome==1.3.0.post0
packaging==23.2
pages==0.3
piglet-templates==1.3.0
pluggy==1.3.0
pycparser==2.21
pyparsing==3.1.1
PySocks==1.7.1
pytest==7.4.4
selenium==4.16.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.24.0
trio-websocket==0.11.1
urllib3==2.1.0
wsproto==1.2.0
zipp==3.17.0
